+++ flutter_inappwebview/example/integration_test/in_app_webview/main.dart
@@ -89,6 +89,7 @@ part 'request_image_ref.dart';
 part 'resize_webview.dart';
 part 'web_archive.dart';
 part 'set_custom_useragent.dart';
+part 'obscured_content_insets.dart';
 part 'screen_time.dart';
 part 'set_get_settings.dart';
 part 'set_web_contents_debugging_enabled.dart';
@@ -205,6 +206,7 @@ void main() {
     webViewAssetLoader();
     onContentSizeChanged();
     screenTime();
+    obscuredContentInsets();
     keepAlive();
 
     // `printCurrentPage` MUST BE LAST, and must stay last.